requirements.txt
torch>=1.12
torchvision>=0.13
pandas
numpy
Pillow
scikit-learn

hydra-core
tqdm
clearml

rich

pathlib

requests
numba
kaggle
